refactor: move some of the validation tests to new files
