+++ validations/src/validations.rs
@@ -2211,6 +2211,11 @@ impl VrfSlots {
         )
         .unwrap()
     }
+
+    /// Return the target hash for each slot
+    pub fn target_hashes(&self) -> &[Hash] {
+        &self.target_hashes
+    }
 }
 
 /// Function to calculate a merkle tree from a transaction vector
@@ -2479,299 +2484,10 @@ pub fn verify_signatures(
 
 #[cfg(test)]
 mod tests {
-    use witnet_data_structures::{
-        chain::Environment,
-        mainnet_validations::{TapiEngine, SECOND_HARD_FORK},
-        radon_error::RadonError,
-    };
-    use witnet_rad::types::{float::RadonFloat, integer::RadonInteger};
-
     use super::*;
 
     const INITIAL_BLOCK_REWARD: u64 = 250 * 1_000_000_000;
     const HALVING_PERIOD: u32 = 3_500_000;
-
-    // This should only be used in tests
-    fn all_wips_active() -> ActiveWips {
-        let mut tapi_engine = TapiEngine::default();
-        tapi_engine.initialize_wip_information(Environment::Testnet);
-
-        ActiveWips {
-            active_wips: tapi_engine.wip_activation,
-            block_epoch: u32::MAX,
-        }
-    }
-
-    #[test]
-    fn test_compare_candidate_same_section() {
-        let bh_1 = Hash::SHA256([10; 32]);
-        let bh_2 = Hash::SHA256([20; 32]);
-        let rep_1 = Reputation(0);
-        let rep_2 = Reputation(2);
-        let vrf_1 = Hash::SHA256([1; 32]);
-        let vrf_2 = Hash::SHA256([2; 32]);
-        // Only one section and all VRFs are valid
-        let vrf_sections = VrfSlots::default();
-
-        // The candidate with reputation always wins
-        for &bh_i in &[bh_1, bh_2] {
-            for &bh_j in &[bh_1, bh_2] {
-                for &vrf_i in &[vrf_1, vrf_2] {
-                    for &vrf_j in &[vrf_1, vrf_2] {
-                        for &act_i in &[true, false] {
-                            for &act_j in &[true, false] {
-                                assert_eq!(
-                                    compare_block_candidates(
-                                        bh_i,
-                                        rep_1,
-                                        vrf_i,
-                                        act_i,
-                                        bh_j,
-                                        rep_2,
-                                        vrf_j,
-                                        act_j,
-                                        &vrf_sections,
-                                    ),
-                                    Ordering::Less
-                                );
-                                assert_eq!(
-                                    compare_block_candidates(
-                                        bh_i,
-                                        rep_2,
-                                        vrf_i,
-                                        act_i,
-                                        bh_j,
-                                        rep_1,
-                                        vrf_j,
-                                        act_j,
-                                        &vrf_sections,
-                                    ),
-                                    Ordering::Greater
-                                );
-                            }
-                        }
-                    }
-                }
-            }
-        }
-
-        // Equal reputation: the candidate that is active wins
-        for &bh_i in &[bh_1, bh_2] {
-            for &bh_j in &[bh_1, bh_2] {
-                for &vrf_i in &[vrf_1, vrf_2] {
-                    for &vrf_j in &[vrf_1, vrf_2] {
-                        assert_eq!(
-                            compare_block_candidates(
-                                bh_i,
-                                rep_1,
-                                vrf_i,
-                                true,
-                                bh_j,
-                                rep_1,
-                                vrf_j,
-                                false,
-                                &vrf_sections,
-                            ),
-                            Ordering::Greater
-                        );
-                        assert_eq!(
-                            compare_block_candidates(
-                                bh_i,
-                                rep_2,
-                                vrf_i,
-                                false,
-                                bh_j,
-                                rep_2,
-                                vrf_j,
-                                true,
-                                &vrf_sections,
-                            ),
-                            Ordering::Less
-                        );
-                    }
-                }
-            }
-        }
-
-        // Equal reputation and activity: the candidate with lower VRF hash wins
-        for &bh_i in &[bh_1, bh_2] {
-            for &bh_j in &[bh_1, bh_2] {
-                assert_eq!(
-                    compare_block_candidates(
-                        bh_i,
-                        rep_1,
-                        vrf_1,
-                        true,
-                        bh_j,
-                        rep_1,
-                        vrf_2,
-                        true,
-                        &vrf_sections,
-                    ),
-                    Ordering::Greater
-                );
-                assert_eq!(
-                    compare_block_candidates(
-                        bh_i,
-                        rep_1,
-                        vrf_2,
-                        true,
-                        bh_j,
-                        rep_1,
-                        vrf_1,
-                        true,
-                        &vrf_sections,
-                    ),
-                    Ordering::Less
-                );
-            }
-        }
-
-        // Equal reputation, equal activity and equal VRF hash: the candidate with lower block hash wins
-        assert_eq!(
-            compare_block_candidates(
-                bh_1,
-                rep_1,
-                vrf_1,
-                true,
-                bh_2,
-                rep_1,
-                vrf_1,
-                true,
-                &vrf_sections,
-            ),
-            Ordering::Greater
-        );
-        assert_eq!(
-            compare_block_candidates(
-                bh_2,
-                rep_1,
-                vrf_1,
-                true,
-                bh_1,
-                rep_1,
-                vrf_1,
-                true,
-                &vrf_sections,
-            ),
-            Ordering::Less
-        );
-
-        // Everything equal: it is the same block
-        assert_eq!(
-            compare_block_candidates(
-                bh_1,
-                rep_1,
-                vrf_1,
-                true,
-                bh_1,
-                rep_1,
-                vrf_1,
-                true,
-                &vrf_sections,
-            ),
-            Ordering::Equal
-        );
-    }
-
-    #[test]
-    fn test_compare_candidate_different_section() {
-        let bh_1 = Hash::SHA256([10; 32]);
-        let bh_2 = Hash::SHA256([20; 32]);
-        let rep_1 = Reputation(0);
-        let rep_2 = Reputation(2);
-        // Candidate 1 should always be better than candidate 2
-        let vrf_sections = VrfSlots::from_rf(16, 1, 2, 1001, 0, 0, &all_wips_active());
-        // Candidate 1 is in section 0
-        let vrf_1 = vrf_sections.target_hashes[0];
-        // Candidate 2 is in section 1
-        let vrf_2 = vrf_sections.target_hashes[1];
-
-        // The candidate in the lower section always wins
-        for &bh_i in &[bh_1, bh_2] {
-            for &bh_j in &[bh_1, bh_2] {
-                for &rep_i in &[rep_1, rep_2] {
-                    for &rep_j in &[rep_1, rep_2] {
-                        for &act_i in &[true, false] {
-                            for &act_j in &[true, false] {
-                                assert_eq!(
-                                    compare_block_candidates(
-                                        bh_i,
-                                        rep_i,
-                                        vrf_1,
-                                        act_i,
-                                        bh_j,
-                                        rep_j,
-                                        vrf_2,
-                                        act_j,
-                                        &vrf_sections,
-                                    ),
-                                    Ordering::Greater
-                                );
-                                assert_eq!(
-                                    compare_block_candidates(
-                                        bh_i,
-                                        rep_i,
-                                        vrf_2,
-                                        act_i,
-                                        bh_j,
-                                        rep_j,
-                                        vrf_1,
-                                        act_j,
-                                        &vrf_sections,
-                                    ),
-                                    Ordering::Less
-                                );
-                            }
-                        }
-                    }
-                }
-            }
-        }
-    }
-
-    #[test]
-    fn test_compare_candidate_different_reputation_bigger_than_zero() {
-        let bh_1 = Hash::SHA256([10; 32]);
-        let bh_2 = Hash::SHA256([20; 32]);
-        let rep_1 = Reputation(1);
-        let rep_2 = Reputation(2);
-        let vrf_1 = Hash::SHA256([1; 32]);
-        let vrf_2 = Hash::SHA256([2; 32]);
-        // Only one section and all VRFs are valid
-        let vrf_sections = VrfSlots::default();
-
-        // In case of active nodes with reputation, the difference will be the vrf not the reputation
-        assert_eq!(
-            compare_block_candidates(
-                bh_1,
-                rep_1,
-                vrf_1,
-                true,
-                bh_2,
-                rep_2,
-                vrf_2,
-                true,
-                &vrf_sections,
-            ),
-            Ordering::Greater
-        );
-
-        assert_eq!(
-            compare_block_candidates(
-                bh_1,
-                rep_1,
-                vrf_2,
-                true,
-                bh_2,
-                rep_2,
-                vrf_1,
-                true,
-                &vrf_sections,
-            ),
-            Ordering::Less
-        );
-    }
 
     #[test]
     fn wit_decimal_places() {
@@ -2826,194 +2542,6 @@ mod tests {
         assert_eq!(reward(3_500_000 * 100), 0);
     }
 
-    // FIXME: Allow for now, wait for https://github.com/rust-lang/rust/issues/67058 to reach stable
-    #[allow(clippy::cast_possible_truncation)]
-    #[test]
-    fn target_randpoe() {
-        // This test is only valid before the first hard fork
-        let a = ActiveWips {
-            active_wips: Default::default(),
-            block_epoch: 1001,
-        };
-        let rf = 1;
-        let minimum_difficulty = 2000;
-        let max_hash = Hash::with_first_u32(0xFFFF_FFFF);
-        let (t00, p00) = calculate_randpoe_threshold(0, rf, 1001, minimum_difficulty, 0, &a);
-        let (t01, p01) = calculate_randpoe_threshold(1, rf, 1001, minimum_difficulty, 0, &a);
-        assert_eq!(t00, max_hash);
-        assert_eq!(t00, t01);
-        assert_eq!((p00 * 100_f64).round() as i128, 100);
-        assert_eq!(
-            (p00 * 100_f64).round() as i128,
-            (p01 * 100_f64).round() as i128
-        );
-        let (t02, p02) = calculate_randpoe_threshold(2, rf, 1001, minimum_difficulty, 0, &a);
-        assert_eq!(t02, Hash::with_first_u32(0x7FFF_FFFF));
-        assert_eq!((p02 * 100_f64).round() as i128, 50);
-        let (t03, p03) = calculate_randpoe_threshold(3, rf, 1001, minimum_difficulty, 0, &a);
-        assert_eq!(t03, Hash::with_first_u32(0x5555_5555));
-        assert_eq!((p03 * 100_f64).round() as i128, 33);
-        let (t04, p04) = calculate_randpoe_threshold(4, rf, 1001, minimum_difficulty, 0, &a);
-        assert_eq!(t04, Hash::with_first_u32(0x3FFF_FFFF));
-        assert_eq!((p04 * 100_f64).round() as i128, 25);
-        let (t05, p05) = calculate_randpoe_threshold(1024, rf, 1001, minimum_difficulty, 0, &a);
-        assert_eq!(t05, Hash::with_first_u32(0x003F_FFFF));
-        assert_eq!((p05 * 100_f64).round() as i128, 0);
-        let (t06, p06) =
-            calculate_randpoe_threshold(1024 * 1024, rf, 1001, minimum_difficulty, 0, &a);
-        assert_eq!(t06, Hash::with_first_u32(0x0000_0FFF));
-        assert_eq!((p06 * 100_f64).round() as i128, 0);
-    }
-
-    #[allow(clippy::cast_possible_truncation)]
-    #[test]
-    fn target_randpoe_initial_difficulty() {
-        // This test is only valid before the first hard fork
-        let a = ActiveWips {
-            active_wips: Default::default(),
-            block_epoch: 1,
-        };
-        let (t, p) = calculate_randpoe_threshold(2, 1, 1, 4, 10, &a);
-        assert_eq!(t, Hash::with_first_u32(0x3FFF_FFFF));
-        assert_eq!((p * 100_f64).round() as i128, 25);
-
-        let (t, p) = calculate_randpoe_threshold(2, 1, 11, 4, 10, &a);
-        assert_eq!(t, Hash::with_first_u32(0x7FFF_FFFF));
-        assert_eq!((p * 100_f64).round() as i128, 50);
-    }
-
-    #[allow(clippy::cast_possible_truncation, clippy::float_cmp)]
-    #[test]
-    fn target_randpoe_minimum_difficulty() {
-        let replication_factor = 2;
-        let minimum_difficulty = 2000;
-        // Before first hard fork
-        let active_wips = ActiveWips {
-            active_wips: Default::default(),
-            block_epoch: 1,
-        };
-
-        let total_identities = 1000;
-        let expected_prob = (1_f64 / f64::from(total_identities)) * f64::from(replication_factor);
-        assert_eq!(expected_prob, 0.002_f64);
-        let (_, p) = calculate_randpoe_threshold(
-            total_identities,
-            replication_factor,
-            1,
-            minimum_difficulty,
-            0,
-            &active_wips,
-        );
-        assert_eq!((p * 10000_f64).round() / 10000_f64, expected_prob);
-
-        // After second hard fork, minimum probability is used
-        let active_wips = all_wips_active();
-        let block_epoch = SECOND_HARD_FORK + 1;
-        let minimum_expected_prob =
-            (1_f64 / f64::from(minimum_difficulty)) * f64::from(replication_factor);
-        assert_eq!(minimum_expected_prob, 0.001_f64);
-        let (_, p) = calculate_randpoe_threshold(
-            total_identities,
-            replication_factor,
-            block_epoch,
-            minimum_difficulty,
-            0,
-            &active_wips,
-        );
-        assert_eq!((p * 10000_f64).round() / 10000_f64, minimum_expected_prob);
-
-        let total_identities = 1500;
-        let (_, p) = calculate_randpoe_threshold(
-            total_identities,
-            replication_factor,
-            block_epoch,
-            minimum_difficulty,
-            0,
-            &active_wips,
-        );
-        assert_eq!((p * 10000_f64).round() / 10000_f64, minimum_expected_prob);
-
-        let total_identities = minimum_difficulty - 1;
-        let (_, p) = calculate_randpoe_threshold(
-            total_identities,
-            replication_factor,
-            block_epoch,
-            minimum_difficulty,
-            0,
-            &active_wips,
-        );
-        assert_eq!((p * 10000_f64).round() / 10000_f64, minimum_expected_prob);
-
-        // When achieves a number of identities equals to minimum difficulty,
-        // the calculated probability is equals to the minimum
-        let total_identities = minimum_difficulty;
-        let expected_prob = (1_f64 / f64::from(total_identities)) * f64::from(replication_factor);
-        assert_eq!(expected_prob, minimum_expected_prob);
-        let (_, p) = calculate_randpoe_threshold(
-            total_identities,
-            replication_factor,
-            block_epoch,
-            minimum_difficulty,
-            0,
-            &active_wips,
-        );
-        assert_eq!((p * 10000_f64).round() / 10000_f64, expected_prob);
-
-        // After that, the probability starts to decrease
-        let total_identities = minimum_difficulty + 1;
-        let expected_prob = (1_f64 / f64::from(total_identities)) * f64::from(replication_factor);
-        let expected_prob = (expected_prob * 1_000_000_000_f64).round() / 1_000_000_000_f64;
-        assert_eq!(expected_prob, 0.0009995_f64);
-        let (_, p) = calculate_randpoe_threshold(
-            total_identities,
-            replication_factor,
-            block_epoch,
-            minimum_difficulty,
-            0,
-            &active_wips,
-        );
-        assert_eq!(
-            (p * 1_000_000_000_f64).round() / 1_000_000_000_f64,
-            expected_prob
-        );
-    }
-
-    // FIXME: Allow for now, wait for https://github.com/rust-lang/rust/issues/67058 to reach stable
-    #[allow(clippy::cast_possible_truncation)]
-    #[test]
-    fn target_randpoe_rf_4() {
-        let rf = 4;
-        let minimum_difficulty = 2000;
-        let max_hash = Hash::with_first_u32(0xFFFF_FFFF);
-        // This test is only valid before the first hard fork
-        let a = ActiveWips {
-            active_wips: Default::default(),
-            block_epoch: 1001,
-        };
-        let (t00, p00) = calculate_randpoe_threshold(0, rf, 1001, minimum_difficulty, 0, &a);
-        let (t01, p01) = calculate_randpoe_threshold(1, rf, 1001, minimum_difficulty, 0, &a);
-        assert_eq!(t00, max_hash);
-        assert_eq!(t01, max_hash);
-        assert_eq!((p00 * 100_f64).round() as i128, 100);
-        assert_eq!((p01 * 100_f64).round() as i128, 100);
-        let (t02, p02) = calculate_randpoe_threshold(2, rf, 1001, minimum_difficulty, 0, &a);
-        assert_eq!(t02, max_hash);
-        assert_eq!((p02 * 100_f64).round() as i128, 100);
-        let (t03, p03) = calculate_randpoe_threshold(3, rf, 1001, minimum_difficulty, 0, &a);
-        assert_eq!(t03, max_hash);
-        assert_eq!((p03 * 100_f64).round() as i128, 100);
-        let (t04, p04) = calculate_randpoe_threshold(4, rf, 1001, minimum_difficulty, 0, &a);
-        assert_eq!(t04, max_hash);
-        assert_eq!((p04 * 100_f64).round() as i128, 100);
-        let (t05, p05) = calculate_randpoe_threshold(1024, rf, 1001, minimum_difficulty, 0, &a);
-        assert_eq!(t05, Hash::with_first_u32(0x00FF_FFFF));
-        assert_eq!((p05 * 100_f64).round() as i128, 0);
-        let (t06, p06) =
-            calculate_randpoe_threshold(1024 * 1024, rf, 1001, minimum_difficulty, 0, &a);
-        assert_eq!(t06, Hash::with_first_u32(0x0000_3FFF));
-        assert_eq!((p06 * 100_f64).round() as i128, 0);
-    }
-
     #[test]
     fn test_counter() {
         let mut counter = Counter::new(7);
@@ -3037,1438 +2565,5 @@ mod tests {
         counter.increment(6);
         assert_eq!(counter.max_val, 3);
         assert_eq!(counter.max_pos, None);
-    }
-
-    #[test]
-    fn test_tally_precondition_clause_3_ints_vs_1_float() {
-        let rad_int = RadonTypes::Integer(RadonInteger::from(1));
-        let rad_float = RadonTypes::Float(RadonFloat::from(1));
-
-        let rad_rep_int = RadonReport::from_result(Ok(rad_int.clone()), &ReportContext::default());
-        let rad_rep_float = RadonReport::from_result(Ok(rad_float), &ReportContext::default());
-
-        let v = vec![
-            rad_rep_int.clone(),
-            rad_rep_int.clone(),
-            rad_rep_int,
-            rad_rep_float,
-        ];
-        let tally_precondition_clause_result =
-            evaluate_tally_precondition_clause(v, 0.70, 4, &all_wips_active()).unwrap();
-
-        if let TallyPreconditionClauseResult::MajorityOfValues {
-            values,
-            liars,
-            errors,
-        } = tally_precondition_clause_result
-        {
-            assert_eq!(values, vec![rad_int.clone(), rad_int.clone(), rad_int]);
-            assert_eq!(liars, vec![false, false, false, true]);
-            assert_eq!(errors, vec![false, false, false, false]);
-        } else {
-            panic!("The result of the tally precondition clause was not `MajorityOfValues`. It was: {:?}", tally_precondition_clause_result);
-        }
-    }
-
-    #[test]
-    fn test_tally_precondition_clause_full_consensus() {
-        let rad_int = RadonTypes::Integer(RadonInteger::from(1));
-
-        let rad_rep_int = RadonReport::from_result(Ok(rad_int.clone()), &ReportContext::default());
-
-        let v = vec![rad_rep_int.clone(), rad_rep_int];
-        let tally_precondition_clause_result =
-            evaluate_tally_precondition_clause(v, 0.99, 2, &all_wips_active()).unwrap();
-
-        if let TallyPreconditionClauseResult::MajorityOfValues {
-            values,
-            liars,
-            errors,
-        } = tally_precondition_clause_result
-        {
-            assert_eq!(values, vec![rad_int.clone(), rad_int]);
-            assert_eq!(liars, vec![false, false]);
-            assert_eq!(errors, vec![false, false]);
-        } else {
-            panic!("The result of the tally precondition clause was not `MajorityOfValues`. It was: {:?}", tally_precondition_clause_result);
-        }
-    }
-
-    #[test]
-    fn test_tally_precondition_clause_exact_consensus() {
-        let rad_int = RadonTypes::Integer(RadonInteger::from(1));
-
-        let rad_rep_int = RadonReport::from_result(Ok(rad_int.clone()), &ReportContext::default());
-
-        let v = vec![rad_rep_int.clone(), rad_rep_int];
-        let tally_precondition_clause_result =
-            evaluate_tally_precondition_clause(v, 1., 2, &all_wips_active()).unwrap();
-
-        if let TallyPreconditionClauseResult::MajorityOfValues {
-            values,
-            liars,
-            errors,
-        } = tally_precondition_clause_result
-        {
-            assert_eq!(values, vec![rad_int.clone(), rad_int]);
-            assert_eq!(liars, vec![false, false]);
-            assert_eq!(errors, vec![false, false]);
-        } else {
-            panic!("The result of the tally precondition clause was not `MajorityOfValues`. It was: {:?}", tally_precondition_clause_result);
-        }
-    }
-
-    #[test]
-    fn test_tally_precondition_clause_3_ints_vs_1_error() {
-        let rad_int = RadonTypes::Integer(RadonInteger::from(1));
-        let rad_err = RadError::HttpStatus { status_code: 404 };
-
-        let rad_rep_int = RadonReport::from_result(Ok(rad_int.clone()), &ReportContext::default());
-        let rad_rep_err = RadonReport::from_result(Err(rad_err), &ReportContext::default());
-
-        let v = vec![
-            rad_rep_int.clone(),
-            rad_rep_err,
-            rad_rep_int.clone(),
-            rad_rep_int,
-        ];
-        let tally_precondition_clause_result =
-            evaluate_tally_precondition_clause(v, 0.70, 4, &all_wips_active()).unwrap();
-
-        if let TallyPreconditionClauseResult::MajorityOfValues {
-            values,
-            liars,
-            errors,
-        } = tally_precondition_clause_result
-        {
-            assert_eq!(values, vec![rad_int.clone(), rad_int.clone(), rad_int]);
-            assert_eq!(liars, vec![false, true, false, false]);
-            assert_eq!(errors, vec![false, true, false, false]);
-        } else {
-            panic!("The result of the tally precondition clause was not `MajorityOfValues`. It was: {:?}", tally_precondition_clause_result);
-        }
-    }
-
-    #[test]
-    fn test_tally_precondition_clause_majority_of_errors() {
-        let rad_int = RadonTypes::Integer(RadonInteger::from(1));
-        let rad_err = RadonError::try_from(RadError::HttpStatus { status_code: 0 }).unwrap();
-
-        let rad_rep_int = RadonReport::from_result(Ok(rad_int), &ReportContext::default());
-        let rad_rep_err = RadonReport::from_result(
-            Ok(RadonTypes::RadonError(rad_err.clone())),
-            &ReportContext::default(),
-        );
-
-        let v = vec![
-            rad_rep_err.clone(),
-            rad_rep_err.clone(),
-            rad_rep_err,
-            rad_rep_int,
-        ];
-        let tally_precondition_clause_result =
-            evaluate_tally_precondition_clause(v, 0.70, 4, &all_wips_active()).unwrap();
-
-        if let TallyPreconditionClauseResult::MajorityOfErrors { errors_mode } =
-            tally_precondition_clause_result
-        {
-            assert_eq!(errors_mode, rad_err);
-        } else {
-            panic!("The result of the tally precondition clause was not `MajorityOfErrors`. It was: {:?}", tally_precondition_clause_result);
-        }
-    }
-
-    #[test]
-    fn test_tally_precondition_clause_mode_tie() {
-        let rad_int = RadonTypes::Integer(RadonInteger::from(1));
-        let rad_float = RadonTypes::Float(RadonFloat::from(1));
-
-        let rad_rep_int = RadonReport::from_result(Ok(rad_int), &ReportContext::default());
-        let rad_rep_float = RadonReport::from_result(Ok(rad_float), &ReportContext::default());
-
-        let v = vec![
-            rad_rep_float.clone(),
-            rad_rep_int.clone(),
-            rad_rep_float,
-            rad_rep_int,
-        ];
-        let out =
-            evaluate_tally_precondition_clause(v.clone(), 0.49, 4, &all_wips_active()).unwrap_err();
-
-        assert_eq!(
-            out,
-            RadError::ModeTie {
-                values: RadonArray::from(
-                    v.into_iter()
-                        .map(RadonReport::into_inner)
-                        .collect::<Vec<RadonTypes>>()
-                ),
-                max_count: 2,
-            }
-        );
-    }
-
-    #[test]
-    fn test_tally_precondition_clause_3_errors_vs_2_ints_and_2_floats() {
-        let rad_int = RadonTypes::Integer(RadonInteger::from(1));
-        let rad_float = RadonTypes::Float(RadonFloat::from(1));
-        let rad_err = RadonError::try_from(RadError::HttpStatus { status_code: 0 }).unwrap();
-
-        let rad_rep_int = RadonReport::from_result(Ok(rad_int), &ReportContext::default());
-        let rad_rep_float = RadonReport::from_result(Ok(rad_float), &ReportContext::default());
-        let rad_rep_err = RadonReport::from_result(
-            Ok(RadonTypes::RadonError(rad_err.clone())),
-            &ReportContext::default(),
-        );
-
-        let v = vec![
-            rad_rep_err.clone(),
-            rad_rep_err.clone(),
-            rad_rep_err,
-            rad_rep_float.clone(),
-            rad_rep_int.clone(),
-            rad_rep_float,
-            rad_rep_int,
-        ];
-        let tally_precondition_clause_result =
-            evaluate_tally_precondition_clause(v, 0.40, 7, &all_wips_active()).unwrap();
-
-        if let TallyPreconditionClauseResult::MajorityOfErrors { errors_mode } =
-            tally_precondition_clause_result
-        {
-            assert_eq!(errors_mode, rad_err);
-        } else {
-            panic!("The result of the tally precondition clause was not `MajorityOfErrors`. It was: {:?}", tally_precondition_clause_result);
-        }
-    }
-
-    #[test]
-    fn test_tally_precondition_clause_no_commits() {
-        let v = vec![];
-        let out = evaluate_tally_precondition_clause(v, 0.51, 0, &all_wips_active()).unwrap_err();
-
-        assert_eq!(out, RadError::InsufficientCommits);
-    }
-
-    #[test]
-    fn test_tally_precondition_clause_no_reveals() {
-        let v = vec![];
-        let out = evaluate_tally_precondition_clause(v, 0.51, 1, &all_wips_active()).unwrap_err();
-
-        assert_eq!(out, RadError::NoReveals);
-    }
-
-    #[test]
-    fn test_tally_precondition_clause_all_errors() {
-        let rad_err = RadonError::try_from(RadError::HttpStatus { status_code: 0 }).unwrap();
-        let rad_rep_err = RadonReport::from_result(
-            Ok(RadonTypes::RadonError(rad_err.clone())),
-            &ReportContext::default(),
-        );
-
-        let v = vec![
-            rad_rep_err.clone(),
-            rad_rep_err.clone(),
-            rad_rep_err.clone(),
-            rad_rep_err,
-        ];
-        let tally_precondition_clause_result =
-            evaluate_tally_precondition_clause(v, 0.51, 4, &all_wips_active()).unwrap();
-
-        if let TallyPreconditionClauseResult::MajorityOfErrors { errors_mode } =
-            tally_precondition_clause_result
-        {
-            assert_eq!(errors_mode, rad_err);
-        } else {
-            panic!("The result of the tally precondition clause was not `MajorityOfErrors`. It was: {:?}", tally_precondition_clause_result);
-        }
-    }
-
-    #[test]
-    fn test_tally_precondition_clause_insufficient_consensus() {
-        let rad_int = RadonTypes::Integer(RadonInteger::from(1));
-        let rad_float = RadonTypes::Float(RadonFloat::from(1));
-
-        let rad_rep_int = RadonReport::from_result(Ok(rad_int), &ReportContext::default());
-        let rad_rep_float = RadonReport::from_result(Ok(rad_float), &ReportContext::default());
-
-        let v = vec![
-            rad_rep_float.clone(),
-            rad_rep_int.clone(),
-            rad_rep_float,
-            rad_rep_int,
-        ];
-        let out = evaluate_tally_precondition_clause(v, 0.51, 4, &all_wips_active()).unwrap_err();
-
-        assert_eq!(
-            out,
-            RadError::InsufficientConsensus {
-                achieved: 0.5,
-                required: 0.51,
-            }
-        );
-    }
-
-    #[test]
-    fn test_tally_precondition_clause_errors_insufficient_consensus() {
-        // Two revealers that report two different errors result in `InsufficientConsensus`
-        // because there is only 50% consensus (1/2)
-        let rad_err1 = RadonError::try_from(RadError::HttpStatus { status_code: 0 }).unwrap();
-        let rad_err2 = RadonError::try_from(RadError::RetrieveTimeout).unwrap();
-        let rad_rep_err1 = RadonReport::from_result(
-            Ok(RadonTypes::RadonError(rad_err1)),
-            &ReportContext::default(),
-        );
-        let rad_rep_err2 = RadonReport::from_result(
-            Ok(RadonTypes::RadonError(rad_err2)),
-            &ReportContext::default(),
-        );
-
-        let v = vec![rad_rep_err1, rad_rep_err2];
-        let out = evaluate_tally_precondition_clause(v, 0.51, 2, &all_wips_active()).unwrap_err();
-
-        assert_eq!(
-            out,
-            RadError::InsufficientConsensus {
-                achieved: 0.5,
-                required: 0.51,
-            }
-        );
-    }
-
-    #[test]
-    fn test_tally_precondition_clause_errors_mode_tie() {
-        // Two revealers that report two different errors when min_consensus is below 50%
-        // result in RadError::ModeTie
-        let rad_err1 = RadonError::try_from(RadError::HttpStatus { status_code: 0 }).unwrap();
-        let rad_err2 = RadonError::try_from(RadError::RetrieveTimeout).unwrap();
-        let rad_rep_err1 = RadonReport::from_result(
-            Ok(RadonTypes::RadonError(rad_err1)),
-            &ReportContext::default(),
-        );
-        let rad_rep_err2 = RadonReport::from_result(
-            Ok(RadonTypes::RadonError(rad_err2)),
-            &ReportContext::default(),
-        );
-
-        let v = vec![rad_rep_err1, rad_rep_err2];
-        let out =
-            evaluate_tally_precondition_clause(v.clone(), 0.49, 2, &all_wips_active()).unwrap_err();
-
-        assert_eq!(
-            out,
-            RadError::ModeTie {
-                values: RadonArray::from(
-                    v.into_iter()
-                        .map(RadonReport::into_inner)
-                        .collect::<Vec<RadonTypes>>()
-                ),
-                max_count: 1,
-            }
-        );
-    }
-
-    #[test]
-    fn vrf_sections() {
-        let h0 = Hash::default();
-        let h1 = Hash::with_first_u32(1);
-        let h2 = Hash::with_first_u32(2);
-        let h3 = Hash::with_first_u32(3);
-        let a = VrfSlots::new(vec![]);
-        assert_eq!(a.slot(&h0), 0);
-
-        let a = VrfSlots::new(vec![h0]);
-        assert_eq!(a.slot(&h0), 0);
-        assert_eq!(a.slot(&h1), 1);
-
-        let a = VrfSlots::new(vec![h0, h2]);
-        assert_eq!(a.slot(&h0), 0);
-        assert_eq!(a.slot(&h1), 1);
-        assert_eq!(a.slot(&h2), 1);
-        assert_eq!(a.slot(&h3), 2);
-    }
-}
-
-#[cfg(test)]
-mod reppoe_tests {
-    use super::*;
-    use witnet_data_structures::{
-        chain::{Alpha, Environment},
-        mainnet_validations::TapiEngine,
-    };
-
-    // This should only be used in tests
-    fn all_wips_active() -> ActiveWips {
-        let mut tapi_engine = TapiEngine::default();
-        tapi_engine.initialize_wip_information(Environment::Testnet);
-
-        ActiveWips {
-            active_wips: tapi_engine.wip_activation,
-            block_epoch: u32::MAX,
-        }
-    }
-
-    fn calculate_reppoe_threshold_v1(
-        rep_eng: &ReputationEngine,
-        pkh: &PublicKeyHash,
-        num_witnesses: u16,
-        minimum_difficulty: u32,
-    ) -> (Hash, f64) {
-        let active_wips = ActiveWips {
-            active_wips: HashMap::default(),
-            block_epoch: 0,
-        };
-        assert!(!active_wips.wip0016());
-        assert!(!active_wips.third_hard_fork());
-        calculate_reppoe_threshold(
-            rep_eng,
-            pkh,
-            num_witnesses,
-            minimum_difficulty,
-            &active_wips,
-        )
-    }
-
-    fn calculate_reppoe_threshold_v2(
-        rep_eng: &ReputationEngine,
-        pkh: &PublicKeyHash,
-        num_witnesses: u16,
-        minimum_difficulty: u32,
-    ) -> (Hash, f64) {
-        let mut active_wips = ActiveWips {
-            active_wips: HashMap::default(),
-            block_epoch: 0,
-        };
-        active_wips
-            .active_wips
-            .insert("THIRD_HARD_FORK".to_string(), 0);
-        assert!(!active_wips.wip0016());
-        assert!(active_wips.third_hard_fork());
-        calculate_reppoe_threshold(
-            rep_eng,
-            pkh,
-            num_witnesses,
-            minimum_difficulty,
-            &active_wips,
-        )
-    }
-
-    fn calculate_reppoe_threshold_v3(
-        rep_eng: &ReputationEngine,
-        pkh: &PublicKeyHash,
-        num_witnesses: u16,
-        minimum_difficulty: u32,
-    ) -> (Hash, f64) {
-        let mut active_wips = all_wips_active();
-        active_wips
-            .active_wips
-            .insert("WIP0014-0016".to_string(), 0);
-        assert!(active_wips.wip0016());
-        assert!(active_wips.third_hard_fork());
-        calculate_reppoe_threshold(
-            rep_eng,
-            pkh,
-            num_witnesses,
-            minimum_difficulty,
-            &active_wips,
-        )
-    }
-
-    // Auxiliar function to add reputation
-    fn add_rep(rep_engine: &mut ReputationEngine, alpha: u32, pkh: PublicKeyHash, rep: u32) {
-        rep_engine
-            .trs_mut()
-            .gain(Alpha(alpha), vec![(pkh, Reputation(rep))])
-            .unwrap();
-    }
-
-    // FIXME: Allow for now, wait for https://github.com/rust-lang/rust/issues/67058 to reach stable
-    #[allow(clippy::cast_possible_truncation)]
-    #[test]
-    fn target_reppoe_v1() {
-        let mut rep_engine = ReputationEngine::new(1000);
-        let id1 = PublicKeyHash::from_bytes(&[1; 20]).unwrap();
-        add_rep(&mut rep_engine, 10, id1, 50);
-        rep_engine.ars_mut().push_activity(vec![id1]);
-
-        // 100% when we have all the reputation
-        let (t00, p00) = calculate_reppoe_threshold_v1(&rep_engine, &id1, 1, 2000);
-        assert_eq!(t00, Hash::with_first_u32(0xFFFF_FFFF));
-        assert_eq!((p00 * 100_f64).round() as i128, 100);
-
-        let id2 = PublicKeyHash::from_bytes(&[2; 20]).unwrap();
-        add_rep(&mut rep_engine, 10, id2, 50);
-        rep_engine.ars_mut().push_activity(vec![id2]);
-
-        // 50% when there are 2 nodes with 50% of the reputation each
-        let (t01, p01) = calculate_reppoe_threshold_v1(&rep_engine, &id1, 1, 2000);
-        assert_eq!(t01, Hash::with_first_u32(0x7FFF_FFFF));
-        assert_eq!((p01 * 100_f64).round() as i128, 50);
-    }
-
-    // FIXME: Allow for now, wait for https://github.com/rust-lang/rust/issues/67058 to reach stable
-    #[allow(clippy::cast_possible_truncation)]
-    #[test]
-    fn target_reppoe_v2() {
-        let mut rep_engine = ReputationEngine::new(1000);
-        let id1 = PublicKeyHash::from_bytes(&[1; 20]).unwrap();
-        add_rep(&mut rep_engine, 10, id1, 50);
-        rep_engine.ars_mut().push_activity(vec![id1]);
-
-        // 0.05% when the total reputation is less than 2000
-        let (t00, p00) = calculate_reppoe_threshold_v2(&rep_engine, &id1, 1, 2000);
-        assert_eq!(t00, Hash::with_first_u32(0xFFFF_FFFF / 2000));
-        // 1 / 2000 = 500 / 1_000_000
-        assert_eq!((p00 * 1_000_000_f64).round() as i128, 500);
-
-        let id2 = PublicKeyHash::from_bytes(&[2; 20]).unwrap();
-        add_rep(&mut rep_engine, 10, id2, 50);
-        rep_engine.ars_mut().push_activity(vec![id2]);
-
-        // 0.05% when the total reputation is less than 2000
-        let (t01, p01) = calculate_reppoe_threshold_v2(&rep_engine, &id1, 1, 2000);
-        assert_eq!(t01, Hash::with_first_u32(0xFFFF_FFFF / 2000));
-        // 1 / 2000 = 500 / 1_000_000
-        assert_eq!((p01 * 1_000_000_f64).round() as i128, 500);
-    }
-
-    // FIXME: Allow for now, wait for https://github.com/rust-lang/rust/issues/67058 to reach stable
-    #[allow(clippy::cast_possible_truncation)]
-    #[test]
-    fn target_reppoe_v3() {
-        let mut rep_engine = ReputationEngine::new(1000);
-        let id1 = PublicKeyHash::from_bytes(&[1; 20]).unwrap();
-        add_rep(&mut rep_engine, 10, id1, 50);
-        rep_engine.ars_mut().push_activity(vec![id1]);
-
-        // 0.05% * 51 when the total reputation is less than 2000
-        // 51 is the reputation of this node (50 + 1 because active)
-        let (t00, p00) = calculate_reppoe_threshold_v3(&rep_engine, &id1, 1, 2000);
-        // 0xFFFF_FFFF / 2000 * 51
-        assert_eq!(t00, Hash::with_first_u32(0x06872b02));
-        assert_eq!((p00 * 100_f64).round() as i128, 3);
-
-        let id2 = PublicKeyHash::from_bytes(&[2; 20]).unwrap();
-        add_rep(&mut rep_engine, 10, id2, 50);
-        rep_engine.ars_mut().push_activity(vec![id2]);
-
-        // 0.05% * 51 when the total reputation is less than 2000
-        // 51 is the reputation of this node (50 + 1 because active)
-        let (t01, p01) = calculate_reppoe_threshold_v3(&rep_engine, &id1, 1, 2000);
-        // 0xFFFF_FFFF / 2000 * 51
-        assert_eq!(t01, Hash::with_first_u32(0x06872b02));
-        assert_eq!((p01 * 100_f64).round() as i128, 3);
-    }
-
-    #[test]
-    // FIXME: Allow for now, wait for https://github.com/rust-lang/rust/issues/67058 to reach stable
-    #[allow(
-        clippy::cast_possible_truncation,
-        clippy::cognitive_complexity,
-        clippy::cast_sign_loss
-    )]
-    fn target_reppoe_specific_example_v1() {
-        let mut rep_engine = ReputationEngine::new(1000);
-        let mut ids = vec![];
-        for i in 0..8 {
-            ids.push(PublicKeyHash::from_bytes(&[i; 20]).unwrap());
-        }
-        rep_engine.ars_mut().push_activity(ids.clone());
-
-        add_rep(&mut rep_engine, 10, ids[0], 79);
-        add_rep(&mut rep_engine, 10, ids[1], 9);
-        add_rep(&mut rep_engine, 10, ids[2], 1);
-        add_rep(&mut rep_engine, 10, ids[3], 1);
-        add_rep(&mut rep_engine, 10, ids[4], 1);
-        add_rep(&mut rep_engine, 10, ids[5], 1);
-
-        let rep_thresholds = |thres| {
-            let mut v = vec![];
-            for id in ids.iter() {
-                v.push(
-                    (calculate_reppoe_threshold_v1(&rep_engine, id, thres, 2000).1 * 1_000_000_f64)
-                        .round() as u32,
-                );
-            }
-            v
-        };
-
-        assert_eq!(
-            rep_thresholds(1),
-            vec![280_000, 230_000, 190_000, 50_000, 140_000, 90_000, 10_000, 10_000]
-        );
-        assert_eq!(
-            rep_thresholds(2),
-            vec![560_000, 460_000, 380_000, 100_000, 280_000, 180_000, 20_000, 20_000]
-        );
-        assert_eq!(
-            rep_thresholds(3),
-            vec![840_000, 690_000, 570_000, 150_000, 420_000, 270_000, 30_000, 30_000]
-        );
-        assert_eq!(
-            rep_thresholds(4),
-            vec![1_000_000, 1_000_000, 950_000, 250_000, 700_000, 450_000, 50_000, 50_000]
-        );
-        assert_eq!(
-            rep_thresholds(5),
-            vec![1_000_000, 1_000_000, 1_000_000, 350_000, 980_000, 630_000, 70_000, 70_000]
-        );
-        assert_eq!(
-            rep_thresholds(6),
-            vec![1_000_000, 1_000_000, 1_000_000, 750_000, 1_000_000, 1_000_000, 150_000, 150_000]
-        );
-        assert_eq!(
-            rep_thresholds(7),
-            vec![
-                1_000_000, 1_000_000, 1_000_000, 1_000_000, 1_000_000, 1_000_000, 500_000, 500_000
-            ]
-        );
-        assert_eq!(rep_thresholds(8), vec![1_000_000; 8]);
-        assert_eq!(rep_thresholds(9), vec![1_000_000; 8]);
-        assert_eq!(rep_thresholds(10), vec![1_000_000; 8]);
-    }
-
-    #[test]
-    // FIXME: Allow for now, wait for https://github.com/rust-lang/rust/issues/67058 to reach stable
-    #[allow(
-        clippy::cast_possible_truncation,
-        clippy::cognitive_complexity,
-        clippy::cast_sign_loss
-    )]
-    fn target_reppoe_specific_example_v2() {
-        let mut rep_engine = ReputationEngine::new(1000);
-        let mut ids = vec![];
-        for i in 0..8 {
-            ids.push(PublicKeyHash::from_bytes(&[i; 20]).unwrap());
-        }
-        rep_engine.ars_mut().push_activity(ids.clone());
-
-        add_rep(&mut rep_engine, 10, ids[0], 79);
-        add_rep(&mut rep_engine, 10, ids[1], 9);
-        add_rep(&mut rep_engine, 10, ids[2], 1);
-        add_rep(&mut rep_engine, 10, ids[3], 1);
-        add_rep(&mut rep_engine, 10, ids[4], 1);
-        add_rep(&mut rep_engine, 10, ids[5], 1);
-
-        let rep_thresholds = |thres| {
-            let mut v = vec![];
-            for id in ids.iter() {
-                v.push(
-                    (calculate_reppoe_threshold_v2(&rep_engine, id, thres, 2000).1 * 1_000_000_f64)
-                        .round() as u32,
-                );
-            }
-            v
-        };
-
-        assert_eq!(rep_thresholds(1), vec![500; 8]);
-        assert_eq!(rep_thresholds(2), vec![1000; 8]);
-        assert_eq!(rep_thresholds(3), vec![1500; 8]);
-        assert_eq!(rep_thresholds(4), vec![2500; 8]);
-        assert_eq!(rep_thresholds(5), vec![3500; 8]);
-        assert_eq!(rep_thresholds(6), vec![7500; 8]);
-        assert_eq!(rep_thresholds(7), vec![25_000; 8]);
-        assert_eq!(rep_thresholds(8), vec![50_000; 8]);
-        assert_eq!(rep_thresholds(9), vec![1_000_000; 8]);
-        assert_eq!(rep_thresholds(10), vec![1_000_000; 8]);
-    }
-
-    #[test]
-    // FIXME: Allow for now, wait for https://github.com/rust-lang/rust/issues/67058 to reach stable
-    #[allow(
-        clippy::cast_possible_truncation,
-        clippy::cognitive_complexity,
-        clippy::cast_sign_loss
-    )]
-    fn target_reppoe_specific_example_v3() {
-        let mut rep_engine = ReputationEngine::new(1000);
-        let mut ids = vec![];
-        for i in 0..8 {
-            ids.push(PublicKeyHash::from_bytes(&[i; 20]).unwrap());
-        }
-        rep_engine.ars_mut().push_activity(ids.clone());
-
-        add_rep(&mut rep_engine, 10, ids[0], 79);
-        add_rep(&mut rep_engine, 10, ids[1], 9);
-        add_rep(&mut rep_engine, 10, ids[2], 1);
-        add_rep(&mut rep_engine, 10, ids[3], 1);
-        add_rep(&mut rep_engine, 10, ids[4], 1);
-        add_rep(&mut rep_engine, 10, ids[5], 1);
-
-        let rep_thresholds = |thres| {
-            let mut v = vec![];
-            for id in ids.iter() {
-                v.push(
-                    (calculate_reppoe_threshold_v3(&rep_engine, id, thres, 2000).1 * 1_000_000_f64)
-                        .round() as u32,
-                );
-            }
-            v
-        };
-
-        assert_eq!(
-            rep_thresholds(1),
-            vec![14_000, 11_500, 9500, 2500, 7000, 4500, 500, 500]
-        );
-        assert_eq!(
-            rep_thresholds(2),
-            vec![28_000, 23_000, 19_000, 5000, 14_000, 9000, 1000, 1000]
-        );
-        assert_eq!(
-            rep_thresholds(3),
-            vec![42_000, 34_500, 28_500, 7500, 21_000, 13_500, 1500, 1500]
-        );
-        assert_eq!(
-            rep_thresholds(4),
-            vec![56_000, 46_000, 38_000, 10_000, 28_000, 18_000, 2000, 2000]
-        );
-        assert_eq!(
-            rep_thresholds(5),
-            vec![70_000, 57_500, 47_500, 12_500, 35_000, 22_500, 2500, 2500]
-        );
-        assert_eq!(
-            rep_thresholds(6),
-            vec![84_000, 69_000, 57_000, 15_000, 42_000, 27_000, 3000, 3000]
-        );
-        assert_eq!(
-            rep_thresholds(7),
-            vec![98_000, 80_500, 66_500, 17_500, 49_000, 31_500, 3500, 3500]
-        );
-        assert_eq!(
-            rep_thresholds(8),
-            vec![112_000, 92000, 76_000, 20_000, 56_000, 36_000, 4000, 4000]
-        );
-        assert_eq!(
-            rep_thresholds(9),
-            vec![126_000, 103_500, 85_500, 22_500, 63_000, 40_500, 4500, 4500]
-        );
-        assert_eq!(
-            rep_thresholds(10),
-            vec![140_000, 115_000, 95_000, 25_000, 70_000, 45_000, 5000, 5000]
-        );
-    }
-
-    // FIXME: Allow for now, wait for https://github.com/rust-lang/rust/issues/67058 to reach stable
-    #[allow(clippy::cast_possible_truncation)]
-    #[test]
-    fn target_reppoe_zero_reputation_v1() {
-        // Test the behavior of the algorithm when our node has 0 reputation
-        let mut rep_engine = ReputationEngine::new(1000);
-        let id0 = PublicKeyHash::from_bytes(&[0; 20]).unwrap();
-
-        // 100% when the total reputation is 0
-        let (t00, p00) = calculate_reppoe_threshold_v1(&rep_engine, &id0, 1, 2000);
-        assert_eq!((p00 * 100_f64).round() as i128, 100);
-        assert_eq!(t00, Hash::with_first_u32(0xFFFF_FFFF));
-        let (t01, p01) = calculate_reppoe_threshold_v1(&rep_engine, &id0, 100, 2000);
-        assert_eq!(t01, Hash::with_first_u32(0xFFFF_FFFF));
-        assert_eq!((p01 * 100_f64).round() as i128, 100);
-
-        let id1 = PublicKeyHash::from_bytes(&[1; 20]).unwrap();
-        rep_engine.ars_mut().push_activity(vec![id1]);
-        let (t02, p02) = calculate_reppoe_threshold_v1(&rep_engine, &id0, 1, 2000);
-        assert_eq!(t02, Hash::with_first_u32(0xFFFF_FFFF));
-        assert_eq!((p02 * 100_f64).round() as i128, 100);
-        let (t02b, p02b) = calculate_reppoe_threshold_v1(&rep_engine, &id0, 2, 2000);
-        assert_eq!(t02b, Hash::with_first_u32(0xFFFF_FFFF));
-        assert_eq!((p02b * 100_f64).round() as i128, 100);
-
-        // 50% when the total reputation is 1
-        add_rep(&mut rep_engine, 10, id1, 1);
-        let (t03, p03) = calculate_reppoe_threshold_v1(&rep_engine, &id0, 1, 2000);
-        assert_eq!(t03, Hash::with_first_u32(0x7FFF_FFFF));
-        assert_eq!((p03 * 100_f64).round() as i128, 50);
-        // 100% when the total reputation is 1
-        // but the number of witnesses is greater than the number of active identities
-        let (t03b, p03b) = calculate_reppoe_threshold_v1(&rep_engine, &id0, 2, 2000);
-        assert_eq!(t03b, Hash::with_first_u32(0xFFFF_FFFF));
-        assert_eq!((p03b * 100_f64).round() as i128, 100);
-
-        // 33% when the total reputation is 1 but there are 2 active identities
-        let id2 = PublicKeyHash::from_bytes(&[2; 20]).unwrap();
-        rep_engine.ars_mut().push_activity(vec![id2]);
-        let (t04, p04) = calculate_reppoe_threshold_v1(&rep_engine, &id0, 1, 2000);
-        assert_eq!(t04, Hash::with_first_u32(0x5555_5555));
-        assert_eq!((p04 * 100_f64).round() as i128, 33);
-        // 100% when the total reputation is 1 but there are 2 active identities
-        // but the number of witnesses is greater than the number of active identities
-        let (t04b, p04b) = calculate_reppoe_threshold_v1(&rep_engine, &id0, 2, 2000);
-        assert_eq!(t04b, Hash::with_first_u32(0xFFFF_FFFF));
-        assert_eq!((p04b * 100_f64).round() as i128, 100);
-        // 100% when the total reputation is 1 but there are 2 active identities
-        // but the number of witnesses is greater than the number of active identities
-        let (t04c, p04c) = calculate_reppoe_threshold_v1(&rep_engine, &id0, 3, 2000);
-        assert_eq!(t04c, Hash::with_first_u32(0xFFFF_FFFF));
-        assert_eq!((p04c * 100_f64).round() as i128, 100);
-
-        // 10 identities with 100 total reputation: 1 / (100 + 10) ~= 0.9%
-        for i in 3..=10 {
-            rep_engine
-                .ars_mut()
-                .push_activity(vec![PublicKeyHash::from_bytes(&[i; 20]).unwrap()]);
-        }
-        add_rep(&mut rep_engine, 10, id1, 99);
-        let (t05, p05) = calculate_reppoe_threshold_v1(&rep_engine, &id0, 1, 2000);
-        assert_eq!(t05, Hash::with_first_u32(0x0253_C825));
-        assert_eq!((p05 * 100_f64).round() as i128, 1);
-
-        // 10 identities with 10000 total reputation: 1 / (10000 + 10) ~= 0.01%
-        add_rep(&mut rep_engine, 10, id1, 9900);
-        let (t05b, p05b) = calculate_reppoe_threshold_v2(&rep_engine, &id0, 1, 2000);
-        assert_eq!(t05b, Hash::with_first_u32(0xFFFF_FFFF / (10_000 + 10)));
-        assert_eq!((p05b * 1_000_000_f64).round() as i128, 100);
-    }
-
-    // FIXME: Allow for now, wait for https://github.com/rust-lang/rust/issues/67058 to reach stable
-    #[allow(clippy::cast_possible_truncation)]
-    #[test]
-    fn target_reppoe_zero_reputation_v2() {
-        // Test the behavior of the algorithm when our node has 0 reputation
-        let mut rep_engine = ReputationEngine::new(1000);
-        let id0 = PublicKeyHash::from_bytes(&[0; 20]).unwrap();
-
-        // 100% when the total reputation is 0
-        let (t00, p00) = calculate_reppoe_threshold_v2(&rep_engine, &id0, 1, 2000);
-        assert_eq!(t00, Hash::with_first_u32(0xFFFF_FFFF));
-        assert_eq!((p00 * 100_f64).round() as i128, 100);
-        let (t01, p01) = calculate_reppoe_threshold_v2(&rep_engine, &id0, 100, 2000);
-        assert_eq!(t01, Hash::with_first_u32(0xFFFF_FFFF));
-        assert_eq!((p01 * 100_f64).round() as i128, 100);
-
-        // 0.05% when the total reputation is 0 but there is 1 active identity
-        let id1 = PublicKeyHash::from_bytes(&[1; 20]).unwrap();
-        rep_engine.ars_mut().push_activity(vec![id1]);
-        let (t02, p02) = calculate_reppoe_threshold_v2(&rep_engine, &id0, 1, 2000);
-        assert_eq!(t02, Hash::with_first_u32(0xFFFF_FFFF / 2000));
-        assert_eq!((p02 * 1_000_000_f64).round() as i128, 500);
-        // 100% when the total reputation is 0 and there is 1 active identity,
-        // but the number of witnesses is greater than the number of active identities
-        let (t02b, p02b) = calculate_reppoe_threshold_v2(&rep_engine, &id0, 2, 2000);
-        assert_eq!(t02b, Hash::with_first_u32(0xFFFF_FFFF));
-        assert_eq!((p02b * 100_f64).round() as i128, 100);
-
-        // 0.05% when the total reputation is 1
-        add_rep(&mut rep_engine, 10, id1, 1);
-        let (t03, p03) = calculate_reppoe_threshold_v2(&rep_engine, &id0, 1, 2000);
-        assert_eq!(t03, Hash::with_first_u32(0xFFFF_FFFF / 2000));
-        assert_eq!((p03 * 1_000_000_f64).round() as i128, 500);
-        // 100% when the total reputation is 1
-        // but the number of witnesses is greater than the number of active identities
-        let (t03b, p03b) = calculate_reppoe_threshold_v2(&rep_engine, &id0, 2, 2000);
-        assert_eq!(t03b, Hash::with_first_u32(0xFFFF_FFFF));
-        assert_eq!((p03b * 100_f64).round() as i128, 100);
-
-        // 0.05% when the total reputation is 1 but there are 2 active identities
-        let id2 = PublicKeyHash::from_bytes(&[2; 20]).unwrap();
-        rep_engine.ars_mut().push_activity(vec![id2]);
-        let (t04, p04) = calculate_reppoe_threshold_v2(&rep_engine, &id0, 1, 2000);
-        assert_eq!(t04, Hash::with_first_u32(0xFFFF_FFFF / 2000));
-        assert_eq!((p04 * 1_000_000_f64).round() as i128, 500);
-        // 0.15% when the total reputation is 1 but there are 2 active identities
-        // but the number of witnesses is greater than 1
-        let (t04b, p04b) = calculate_reppoe_threshold_v2(&rep_engine, &id0, 2, 2000);
-        // 0xFFFF_FFFF / 2000 * 3
-        assert_eq!(t04b, Hash::with_first_u32(0x00624dd2));
-        assert_eq!((p04b * 1_000_000_f64).round() as i128, 1500);
-        // 100% when the total reputation is 1 but there are 2 active identities
-        // but the number of witnesses is greater than the number of active identities
-        let (t04c, p04c) = calculate_reppoe_threshold_v2(&rep_engine, &id0, 3, 2000);
-        assert_eq!(t04c, Hash::with_first_u32(0xFFFF_FFFF));
-        assert_eq!((p04c * 100_f64).round() as i128, 100);
-
-        // 10 identities with 100 total reputation: 0.15% (100 is less than the minimum of 2000)
-        for i in 3..=10 {
-            rep_engine
-                .ars_mut()
-                .push_activity(vec![PublicKeyHash::from_bytes(&[i; 20]).unwrap()]);
-        }
-        add_rep(&mut rep_engine, 10, id1, 99);
-        let (t05, p05) = calculate_reppoe_threshold_v2(&rep_engine, &id0, 1, 2000);
-        assert_eq!(t05, Hash::with_first_u32(0xFFFF_FFFF / 2000));
-        assert_eq!((p05 * 1_000_000_f64).round() as i128, 500);
-
-        // 10 identities with 10000 total reputation: 1 / (10000 + 10) ~= 0.01%
-        add_rep(&mut rep_engine, 10, id1, 9900);
-        let (t05b, p05b) = calculate_reppoe_threshold_v2(&rep_engine, &id0, 1, 2000);
-        assert_eq!(t05b, Hash::with_first_u32(0xFFFF_FFFF / (10_000 + 10)));
-        assert_eq!((p05b * 1_000_000_f64).round() as i128, 100);
-    }
-
-    // FIXME: Allow for now, wait for https://github.com/rust-lang/rust/issues/67058 to reach stable
-    #[allow(clippy::cast_possible_truncation)]
-    #[test]
-    fn target_reppoe_zero_reputation_v3() {
-        // Test the behavior of the algorithm when our node has 0 reputation
-        let mut rep_engine = ReputationEngine::new(1000);
-        let id0 = PublicKeyHash::from_bytes(&[0; 20]).unwrap();
-
-        // 0.05% when the total reputation is 0
-        let (t00, p00) = calculate_reppoe_threshold_v3(&rep_engine, &id0, 1, 2000);
-        assert_eq!(t00, Hash::with_first_u32(0xFFFF_FFFF / 2000));
-        assert_eq!((p00 * 1_000_000_f64).round() as i128, 500);
-        // 5% when the total reputation is 0 but the number of witnesses is 100
-        let (t01, p01) = calculate_reppoe_threshold_v3(&rep_engine, &id0, 100, 2000);
-        assert_eq!(t01, Hash::with_first_u32(0xFFFF_FFFF / (2000 / 100)));
-        assert_eq!((p01 * 100_f64).round() as i128, 5);
-
-        let id1 = PublicKeyHash::from_bytes(&[1; 20]).unwrap();
-        rep_engine.ars_mut().push_activity(vec![id1]);
-        let (t02, p02) = calculate_reppoe_threshold_v3(&rep_engine, &id0, 1, 2000);
-        assert_eq!(t02, Hash::with_first_u32(0xFFFF_FFFF / 2000));
-        assert_eq!((p02 * 1_000_000_f64).round() as i128, 500);
-        // 0.10% when the total reputation is 0 and there is 1 active identity,
-        // but the number of witnesses is 2
-        let (t02b, p02b) = calculate_reppoe_threshold_v3(&rep_engine, &id0, 2, 2000);
-        // 0xFFFF_FFFF / 2000 * 2
-        assert_eq!(t02b, Hash::with_first_u32(0x00418937));
-        assert_eq!((p02b * 1_000_000_f64).round() as i128, 1000);
-
-        // 0.05% when the total reputation is 1
-        add_rep(&mut rep_engine, 10, id1, 1);
-        let (t03, p03) = calculate_reppoe_threshold_v3(&rep_engine, &id0, 1, 2000);
-        assert_eq!(t03, Hash::with_first_u32(0xFFFF_FFFF / 2000));
-        assert_eq!((p03 * 1_000_000_f64).round() as i128, 500);
-        // 0.10% when the total reputation is 1
-        // but the number of witnesses is greater than the number of active identities
-        let (t03b, p03b) = calculate_reppoe_threshold_v3(&rep_engine, &id0, 2, 2000);
-        // 0xFFFF_FFFF / 2000 * 2 + 1
-        assert_eq!(t03b, Hash::with_first_u32(0x00418937));
-        assert_eq!((p03b * 1_000_000_f64).round() as i128, 1000);
-
-        // When the total reputation is 1 but there are 2 active identities:
-        // 0.05% when the number of witnesses is 1
-        let id2 = PublicKeyHash::from_bytes(&[2; 20]).unwrap();
-        rep_engine.ars_mut().push_activity(vec![id2]);
-        let (t04, p04) = calculate_reppoe_threshold_v3(&rep_engine, &id0, 1, 2000);
-        assert_eq!(t04, Hash::with_first_u32(0xFFFF_FFFF / 2000));
-        assert_eq!((p04 * 1_000_000_f64).round() as i128, 500);
-        // 0.10% when the number of witnesses is 2
-        let (t04b, p04b) = calculate_reppoe_threshold_v3(&rep_engine, &id0, 2, 2000);
-        // 0xFFFF_FFFF / 2000 * 2
-        assert_eq!(t04b, Hash::with_first_u32(0x00418937));
-        assert_eq!((p04b * 1_000_000_f64).round() as i128, 1000);
-        // 0.15% when the number of witnesses is 3
-        let (t04c, p04c) = calculate_reppoe_threshold_v3(&rep_engine, &id0, 3, 2000);
-        // 0xFFFF_FFFF / 2000 * 3
-        assert_eq!(t04c, Hash::with_first_u32(0x00624dd2));
-        assert_eq!((p04c * 1_000_000_f64).round() as i128, 1500);
-
-        // 10 identities with 100 total reputation: 0.15% (100 is less than the minimum of 2000)
-        for i in 3..=10 {
-            rep_engine
-                .ars_mut()
-                .push_activity(vec![PublicKeyHash::from_bytes(&[i; 20]).unwrap()]);
-        }
-        add_rep(&mut rep_engine, 10, id1, 99);
-        let (t05, p05) = calculate_reppoe_threshold_v3(&rep_engine, &id0, 1, 2000);
-        assert_eq!(t05, Hash::with_first_u32(0xFFFF_FFFF / 2000));
-        assert_eq!((p05 * 1_000_000_f64).round() as i128, 500);
-
-        // 10 identities with 10000 total reputation: 1 / (10000 + 10) ~= 0.01%
-        add_rep(&mut rep_engine, 10, id1, 9900);
-        let (t05b, p05b) = calculate_reppoe_threshold_v3(&rep_engine, &id0, 1, 2000);
-        assert_eq!(t05b, Hash::with_first_u32(0xFFFF_FFFF / (10_000 + 10)));
-        assert_eq!((p05b * 1_000_000_f64).round() as i128, 100);
-    }
-
-    // FIXME: Allow for now, wait for https://github.com/rust-lang/rust/issues/67058 to reach stable
-    #[allow(clippy::cast_possible_truncation)]
-    #[test]
-    fn reppoe_overflow_v1() {
-        // Test the behavior of the algorithm when one node has 99% of the reputation
-        let mut rep_engine = ReputationEngine::new(1000);
-        let id0 = PublicKeyHash::from_bytes(&[0; 20]).unwrap();
-        let id1 = PublicKeyHash::from_bytes(&[1; 20]).unwrap();
-        let id2 = PublicKeyHash::from_bytes(&[2; 20]).unwrap();
-        rep_engine.ars_mut().push_activity(vec![id0]);
-        rep_engine.ars_mut().push_activity(vec![id1]);
-        add_rep(&mut rep_engine, 10, id0, u32::max_value() - 2);
-
-        // Test big values that result in < 100%
-        // Active identity with 100% of the reputation
-        let (t01, p01) = calculate_reppoe_threshold_v1(&rep_engine, &id0, 1, 2000);
-        assert_eq!(t01, Hash::with_first_u32(0xFFFF_FFFE));
-        assert_eq!((p01 * 100_f64).round() as i128, 100);
-        // Active identity with 0 reputation
-        let (t02, p02) = calculate_reppoe_threshold_v1(&rep_engine, &id1, 1, 2000);
-        assert_eq!(t02, Hash::with_first_u32(0x00000001));
-        assert_eq!((p02 * 1_000_000_f64).round() as i128, 0);
-        // Inactive identity with 0 reputation
-        let (t03, p03) = calculate_reppoe_threshold_v1(&rep_engine, &id2, 1, 2000);
-        assert_eq!(t03, Hash::with_first_u32(0x00000001));
-        assert_eq!((p03 * 1_000_000_f64).round() as i128, 0);
-
-        // Repeat test with 2 witnesses
-        // Active identity with 100% of the reputation
-        let (t01, p01) = calculate_reppoe_threshold_v1(&rep_engine, &id0, 2, 2000);
-        // 100% eligibility because the number of witnesses is greater than the number of active identities
-        assert_eq!(t01, Hash::with_first_u32(0xFFFF_FFFF));
-        assert_eq!((p01 * 100_f64).round() as i128, 100);
-        // Active identity with 0 reputation
-        let (t02, p02) = calculate_reppoe_threshold_v1(&rep_engine, &id1, 2, 2000);
-        assert_eq!(t02, Hash::with_first_u32(0xFFFF_FFFF));
-        assert_eq!((p02 * 100_f64).round() as i128, 100);
-        // Inactive identity with 0 reputation
-        let (t03, p03) = calculate_reppoe_threshold_v1(&rep_engine, &id2, 2, 2000);
-        assert_eq!(t03, Hash::with_first_u32(0xFFFF_FFFF));
-        assert_eq!((p03 * 100_f64).round() as i128, 100);
-    }
-
-    // FIXME: Allow for now, wait for https://github.com/rust-lang/rust/issues/67058 to reach stable
-    #[allow(clippy::cast_possible_truncation)]
-    #[test]
-    fn reppoe_overflow_v2() {
-        // Test the behavior of the algorithm when one node has 99% of the reputation
-        let mut rep_engine = ReputationEngine::new(1000);
-        let id0 = PublicKeyHash::from_bytes(&[0; 20]).unwrap();
-        let id1 = PublicKeyHash::from_bytes(&[1; 20]).unwrap();
-        let id2 = PublicKeyHash::from_bytes(&[2; 20]).unwrap();
-        rep_engine.ars_mut().push_activity(vec![id0]);
-        rep_engine.ars_mut().push_activity(vec![id1]);
-        add_rep(&mut rep_engine, 10, id0, u32::max_value() - 2);
-
-        // Test big values that result in < 100%
-        // Active identity with 100% of the reputation
-        let (t01, p01) = calculate_reppoe_threshold_v2(&rep_engine, &id0, 1, 2000);
-        assert_eq!(t01, Hash::with_first_u32(0xFFFF_FFFF / 2000));
-        // 1 / 2000 = 500 / 1_000_000
-        assert_eq!((p01 * 1_000_000_f64).round() as i128, 500);
-        // Active identity with 0 reputation
-        let (t02, p02) = calculate_reppoe_threshold_v2(&rep_engine, &id1, 1, 2000);
-        assert_eq!(t02, Hash::with_first_u32(0x00000001));
-        assert_eq!((p02 * 1_000_000_f64).round() as i128, 0);
-        // Inactive identity with 0 reputation
-        let (t03, p03) = calculate_reppoe_threshold_v2(&rep_engine, &id2, 1, 2000);
-        assert_eq!(t03, Hash::with_first_u32(0x00000001));
-        assert_eq!((p03 * 1_000_000_f64).round() as i128, 0);
-
-        // Repeat test with 2 witnesses
-        // Active identity with 100% of the reputation
-        let (t01, p01) = calculate_reppoe_threshold_v2(&rep_engine, &id0, 2, 2000);
-        // 100% eligibility because the number of witnesses is greater than the number of active identities
-        assert_eq!(t01, Hash::with_first_u32(0xFFFF_FFFF));
-        assert_eq!((p01 * 100_f64).round() as i128, 100);
-        // Active identity with 0 reputation
-        let (t02, p02) = calculate_reppoe_threshold_v2(&rep_engine, &id1, 2, 2000);
-        assert_eq!(t02, Hash::with_first_u32(0xFFFF_FFFF));
-        assert_eq!((p02 * 100_f64).round() as i128, 100);
-        // Inactive identity with 0 reputation
-        let (t03, p03) = calculate_reppoe_threshold_v2(&rep_engine, &id2, 2, 2000);
-        assert_eq!(t03, Hash::with_first_u32(0xFFFF_FFFF));
-        assert_eq!((p03 * 100_f64).round() as i128, 100);
-    }
-
-    // FIXME: Allow for now, wait for https://github.com/rust-lang/rust/issues/67058 to reach stable
-    #[allow(clippy::cast_possible_truncation)]
-    #[test]
-    fn reppoe_overflow_v3() {
-        // Test the behavior of the algorithm when one node has 99% of the reputation
-        let mut rep_engine = ReputationEngine::new(1000);
-        let id0 = PublicKeyHash::from_bytes(&[0; 20]).unwrap();
-        let id1 = PublicKeyHash::from_bytes(&[1; 20]).unwrap();
-        let id2 = PublicKeyHash::from_bytes(&[2; 20]).unwrap();
-        rep_engine.ars_mut().push_activity(vec![id0]);
-        rep_engine.ars_mut().push_activity(vec![id1]);
-        add_rep(&mut rep_engine, 10, id0, u32::max_value() - 2);
-
-        // Test big values that result in < 100%
-        // Active identity with 100% of the reputation
-        let (t01, p01) = calculate_reppoe_threshold_v3(&rep_engine, &id0, 1, 2000);
-        assert_eq!(t01, Hash::with_first_u32(0xFFFF_FFFE));
-        assert_eq!((p01 * 100_f64).round() as i128, 100);
-        // Active identity with 0 reputation
-        let (t02, p02) = calculate_reppoe_threshold_v3(&rep_engine, &id1, 1, 2000);
-        assert_eq!(t02, Hash::with_first_u32(0x00000001));
-        assert_eq!((p02 * 1_000_000_f64).round() as i128, 0);
-        // Inactive identity with 0 reputation
-        let (t03, p03) = calculate_reppoe_threshold_v3(&rep_engine, &id2, 1, 2000);
-        assert_eq!(t03, Hash::with_first_u32(0x00000001));
-        assert_eq!((p03 * 1_000_000_f64).round() as i128, 0);
-
-        // Repeat test with 2 witnesses
-        // Active identity with 100% of the reputation
-        let (t01, p01) = calculate_reppoe_threshold_v3(&rep_engine, &id0, 2, 2000);
-        assert_eq!(t01, Hash::with_first_u32(0xFFFF_FFFF));
-        assert_eq!((p01 * 100_f64).round() as i128, 100);
-        // Active identity with 0 reputation
-        let (t02, p02) = calculate_reppoe_threshold_v3(&rep_engine, &id1, 2, 2000);
-        assert_eq!(t02, Hash::with_first_u32(0x00000002));
-        assert_eq!((p02 * 1_000_000_f64).round() as i128, 0);
-        // Inactive identity with 0 reputation
-        let (t03, p03) = calculate_reppoe_threshold_v3(&rep_engine, &id2, 2, 2000);
-        assert_eq!(t03, Hash::with_first_u32(0x00000002));
-        assert_eq!((p03 * 1_000_000_f64).round() as i128, 0);
-    }
-
-    // FIXME: Allow for now, wait for https://github.com/rust-lang/rust/issues/67058 to reach stable
-    #[allow(clippy::cast_possible_truncation)]
-    #[test]
-    fn reppoe_much_rep_trapezoid_v1() {
-        // Test the behavior of the algorithm when one node has 99% reputation and another node has 1%
-        let mut rep_engine = ReputationEngine::new(1000);
-        let id0 = PublicKeyHash::from_bytes(&[0; 20]).unwrap();
-        let id1 = PublicKeyHash::from_bytes(&[1; 20]).unwrap();
-        let id2 = PublicKeyHash::from_bytes(&[2; 20]).unwrap();
-        rep_engine.ars_mut().push_activity(vec![id0]);
-        rep_engine.ars_mut().push_activity(vec![id1]);
-        add_rep(&mut rep_engine, 10, id0, u32::max_value() - 4);
-        add_rep(&mut rep_engine, 10, id1, 1);
-
-        // Test big values that result in < 100%
-        // Active identity with 99% of the reputation
-        let (t01, p01) = calculate_reppoe_threshold_v1(&rep_engine, &id0, 1, 2000);
-        assert_eq!(t01, Hash::with_first_u32(0xDFFF_FFFE));
-        assert_eq!((p01 * 100_f64).round() as i128, 87);
-        // Active identity with 1 reputation
-        let (t02, p02) = calculate_reppoe_threshold_v1(&rep_engine, &id1, 1, 2000);
-        assert_eq!(t02, Hash::with_first_u32(0x20000001));
-        assert_eq!((p02 * 100_f64).round() as i128, 13);
-        // Inactive identity with 0 reputation
-        let (t03, p03) = calculate_reppoe_threshold_v1(&rep_engine, &id2, 1, 2000);
-        assert_eq!(t03, Hash::with_first_u32(0x00000001));
-        assert_eq!((p03 * 1_000_000_f64).round() as i128, 0);
-
-        // Repeat test with 2 witnesses
-        // Active identity with 99% of the reputation
-        let (t01, p01) = calculate_reppoe_threshold_v1(&rep_engine, &id0, 2, 2000);
-        assert_eq!(t01, Hash::with_first_u32(0xFFFF_FFFF));
-        assert_eq!((p01 * 100_f64).round() as i128, 100);
-        // Active identity with 1 reputation
-        let (t02, p02) = calculate_reppoe_threshold_v1(&rep_engine, &id1, 2, 2000);
-        assert_eq!(t02, Hash::with_first_u32(0xFFFF_FFFF));
-        assert_eq!((p02 * 100_f64).round() as i128, 100);
-        // Inactive identity with 0 reputation
-        let (t03, p03) = calculate_reppoe_threshold_v1(&rep_engine, &id2, 2, 2000);
-        assert_eq!(t03, Hash::with_first_u32(0x00000008));
-        assert_eq!((p03 * 1_000_000_f64).round() as i128, 0);
-    }
-
-    // FIXME: Allow for now, wait for https://github.com/rust-lang/rust/issues/67058 to reach stable
-    #[allow(clippy::cast_possible_truncation)]
-    #[test]
-    fn reppoe_much_rep_trapezoid_v2() {
-        // Test the behavior of the algorithm when one node has 99% reputation and another node has 1%
-        let mut rep_engine = ReputationEngine::new(1000);
-        let id0 = PublicKeyHash::from_bytes(&[0; 20]).unwrap();
-        let id1 = PublicKeyHash::from_bytes(&[1; 20]).unwrap();
-        let id2 = PublicKeyHash::from_bytes(&[2; 20]).unwrap();
-        rep_engine.ars_mut().push_activity(vec![id0]);
-        rep_engine.ars_mut().push_activity(vec![id1]);
-        add_rep(&mut rep_engine, 10, id0, u32::max_value() - 4);
-        add_rep(&mut rep_engine, 10, id1, 1);
-
-        // Test big values that result in < 100%
-        // Active identity with 99% of the reputation
-        let (t01, p01) = calculate_reppoe_threshold_v2(&rep_engine, &id0, 1, 2000);
-        assert_eq!(t01, Hash::with_first_u32(0xFFFF_FFFF / 2000));
-        // 1 / 2000 = 500 / 1_000_000
-        assert_eq!((p01 * 1_000_000_f64).round() as i128, 500);
-        // Active identity with 1 reputation
-        let (t02, p02) = calculate_reppoe_threshold_v2(&rep_engine, &id1, 1, 2000);
-        assert_eq!(t02, Hash::with_first_u32(0xFFFF_FFFF / 2000));
-        // 1 / 2000 = 500 / 1_000_000
-        assert_eq!((p02 * 1_000_000_f64).round() as i128, 500);
-        // Inactive identity with 0 reputation
-        let (t03, p03) = calculate_reppoe_threshold_v2(&rep_engine, &id2, 1, 2000);
-        assert_eq!(t03, Hash::with_first_u32(0x00000001));
-        assert_eq!((p03 * 1_000_000_f64).round() as i128, 0);
-
-        // Repeat test with 2 witnesses
-        // Active identity with 99% of the reputation
-        let (t01, p01) = calculate_reppoe_threshold_v2(&rep_engine, &id0, 2, 2000);
-        assert_eq!(t01, Hash::with_first_u32(0x010624dd));
-        // 8 / 2000 = 4_000 / 1_000_000
-        assert_eq!((p01 * 1_000_000_f64).round() as i128, 4_000);
-        // Active identity with 1 reputation
-        let (t02, p02) = calculate_reppoe_threshold_v2(&rep_engine, &id1, 2, 2000);
-        assert_eq!(t02, Hash::with_first_u32(0x010624dd));
-        // 8 / 2000 = 4_000 / 1_000_000
-        assert_eq!((p02 * 1_000_000_f64).round() as i128, 4_000);
-        // Inactive identity with 0 reputation
-        let (t03, p03) = calculate_reppoe_threshold_v2(&rep_engine, &id2, 2, 2000);
-        assert_eq!(t03, Hash::with_first_u32(0x00000008));
-        assert_eq!((p03 * 1_000_000_f64).round() as i128, 0);
-    }
-
-    // FIXME: Allow for now, wait for https://github.com/rust-lang/rust/issues/67058 to reach stable
-    #[allow(clippy::cast_possible_truncation)]
-    #[test]
-    fn reppoe_much_rep_trapezoid_v3() {
-        // Test the behavior of the algorithm when one node has 99% reputation and another node has 1%
-        let mut rep_engine = ReputationEngine::new(1000);
-        let id0 = PublicKeyHash::from_bytes(&[0; 20]).unwrap();
-        let id1 = PublicKeyHash::from_bytes(&[1; 20]).unwrap();
-        let id2 = PublicKeyHash::from_bytes(&[2; 20]).unwrap();
-        rep_engine.ars_mut().push_activity(vec![id0]);
-        rep_engine.ars_mut().push_activity(vec![id1]);
-        add_rep(&mut rep_engine, 10, id0, u32::max_value() - 4);
-        add_rep(&mut rep_engine, 10, id1, 1);
-
-        // Test big values that result in < 100%
-        // Active identity with 99% of the reputation
-        let (t01, p01) = calculate_reppoe_threshold_v3(&rep_engine, &id0, 1, 2000);
-        assert_eq!(t01, Hash::with_first_u32(0xDFFF_FFFE));
-        assert_eq!((p01 * 100_f64).round() as i128, 87);
-        // Active identity with 1 reputation
-        let (t02, p02) = calculate_reppoe_threshold_v3(&rep_engine, &id1, 1, 2000);
-        assert_eq!(t02, Hash::with_first_u32(0x20000001));
-        assert_eq!((p02 * 100_f64).round() as i128, 13);
-        // Inactive identity with 0 reputation
-        let (t03, p03) = calculate_reppoe_threshold_v3(&rep_engine, &id2, 1, 2000);
-        assert_eq!(t03, Hash::with_first_u32(0x00000001));
-        assert_eq!((p03 * 1_000_000_f64).round() as i128, 0);
-
-        // Repeat test with 2 witnesses
-        // Active identity with 99% of the reputation
-        let (t01, p01) = calculate_reppoe_threshold_v3(&rep_engine, &id0, 2, 2000);
-        assert_eq!(t01, Hash::with_first_u32(0xFFFF_FFFF));
-        assert_eq!((p01 * 100_f64).round() as i128, 100);
-        // Active identity with 1 reputation
-        let (t02, p02) = calculate_reppoe_threshold_v3(&rep_engine, &id1, 2, 2000);
-        assert_eq!(t02, Hash::with_first_u32(0x40000002));
-        assert_eq!((p02 * 100_f64).round() as i128, 25);
-        // Inactive identity with 0 reputation
-        let (t03, p03) = calculate_reppoe_threshold_v3(&rep_engine, &id2, 2, 2000);
-        assert_eq!(t03, Hash::with_first_u32(0x00000002));
-        assert_eq!((p03 * 1_000_000_f64).round() as i128, 0);
-    }
-
-    // FIXME: Allow for now, wait for https://github.com/rust-lang/rust/issues/67058 to reach stable
-    #[allow(clippy::cast_possible_truncation)]
-    #[test]
-    fn reppoe_100_reputed_nodes_v1() {
-        // Test the behavior of the algorithm when 100 nodes have equally large reputation
-        let mut rep_engine = ReputationEngine::new(1000);
-        let gen_id = |i| PublicKeyHash::from_bytes(&[i; 20]).unwrap();
-        let id_rep = gen_id(0);
-        let id_no_active = gen_id(255);
-        rep_engine.ars_mut().push_activity((0..100).map(gen_id));
-        rep_engine
-            .trs_mut()
-            .gain(
-                Alpha(10),
-                (0..100).map(|i| {
-                    let id = gen_id(i);
-                    (id, Reputation(10_000))
-                }),
-            )
-            .unwrap();
-
-        // Active identity with 1% of the reputation
-        let (t01, p01) = calculate_reppoe_threshold_v1(&rep_engine, &id_rep, 1, 2000);
-        assert_eq!(t01, Hash::with_first_u32(0x028f5c28));
-        assert_eq!((p01 * 1_000_000_f64).round() as i128, 10_000);
-        // Inactive identity with 0 reputation
-        let (t02, p02) = calculate_reppoe_threshold_v1(&rep_engine, &id_no_active, 1, 2000);
-        assert_eq!(t02, Hash::with_first_u32(0x000010c6));
-        assert_eq!((p02 * 1_000_000_f64).round() as i128, 1);
-
-        // Repeat test with 2 witnesses
-        // Active identity with 1% of the reputation
-        let (t01, p01) = calculate_reppoe_threshold_v1(&rep_engine, &id_rep, 2, 2000);
-        assert_eq!(t01, Hash::with_first_u32(0x051eb851));
-        assert_eq!((p01 * 1_000_000_f64).round() as i128, 20_000);
-        // Inactive identity with 0 reputation
-        let (t02, p02) = calculate_reppoe_threshold_v1(&rep_engine, &id_no_active, 2, 2000);
-        assert_eq!(t02, Hash::with_first_u32(0x0000218d));
-        assert_eq!((p02 * 1_000_000_f64).round() as i128, 2);
-
-        // Repeat test with 100 witnesses
-        // Active identity with 1% of the reputation
-        let (t01, p01) = calculate_reppoe_threshold_v1(&rep_engine, &id_rep, 100, 2000);
-        assert_eq!(t01, Hash::with_first_u32(0xFFFF_FFFF));
-        assert_eq!((p01 * 100_f64).round() as i128, 100);
-        // Inactive identity with 0 reputation
-        let (t02, p02) = calculate_reppoe_threshold_v1(&rep_engine, &id_no_active, 100, 2000);
-        assert_eq!(t02, Hash::with_first_u32(0x00068d8d));
-        assert_eq!((p02 * 1_000_000_f64).round() as i128, 100);
-
-        // Repeat test with 101 witnesses
-        // Active identity with 1% of the reputation
-        let (t01, p01) = calculate_reppoe_threshold_v1(&rep_engine, &id_rep, 101, 2000);
-        assert_eq!(t01, Hash::with_first_u32(0xFFFF_FFFF));
-        assert_eq!((p01 * 100_f64).round() as i128, 100);
-        // Inactive identity with 0 reputation
-        let (t02, p02) = calculate_reppoe_threshold_v1(&rep_engine, &id_no_active, 101, 2000);
-        assert_eq!(t02, Hash::with_first_u32(0xFFFF_FFFF));
-        assert_eq!((p02 * 100_f64).round() as i128, 100);
-    }
-
-    // FIXME: Allow for now, wait for https://github.com/rust-lang/rust/issues/67058 to reach stable
-    #[allow(clippy::cast_possible_truncation)]
-    #[test]
-    fn reppoe_100_reputed_nodes_v2() {
-        // Test the behavior of the algorithm when 100 nodes have equally large reputation
-        let mut rep_engine = ReputationEngine::new(1000);
-        let gen_id = |i| PublicKeyHash::from_bytes(&[i; 20]).unwrap();
-        let id_rep = gen_id(0);
-        let id_no_active = gen_id(255);
-        rep_engine.ars_mut().push_activity((0..100).map(gen_id));
-        rep_engine
-            .trs_mut()
-            .gain(
-                Alpha(10),
-                (0..100).map(|i| {
-                    let id = gen_id(i);
-                    (id, Reputation(10_000))
-                }),
-            )
-            .unwrap();
-
-        // Active identity with 1% of the reputation
-        let (t01, p01) = calculate_reppoe_threshold_v2(&rep_engine, &id_rep, 1, 2000);
-        assert_eq!(t01, Hash::with_first_u32(0x0020c49b));
-        assert_eq!((p01 * 1_000_000_f64).round() as i128, 500);
-        // Inactive identity with 0 reputation
-        let (t02, p02) = calculate_reppoe_threshold_v2(&rep_engine, &id_no_active, 1, 2000);
-        assert_eq!(t02, Hash::with_first_u32(0x000010c6));
-        assert_eq!((p02 * 1_000_000_f64).round() as i128, 1);
-
-        // Repeat test with 2 witnesses
-        // Active identity with 1% of the reputation
-        let (t01, p01) = calculate_reppoe_threshold_v2(&rep_engine, &id_rep, 2, 2000);
-        assert_eq!(t01, Hash::with_first_u32(0x00418937));
-        assert_eq!((p01 * 1_000_000_f64).round() as i128, 1000);
-        // Inactive identity with 0 reputation
-        let (t02, p02) = calculate_reppoe_threshold_v2(&rep_engine, &id_no_active, 2, 2000);
-        assert_eq!(t02, Hash::with_first_u32(0x0000218d));
-        assert_eq!((p02 * 1_000_000_f64).round() as i128, 2);
-
-        // Repeat test with 100 witnesses
-        // Active identity with 1% of the reputation
-        let (t01, p01) = calculate_reppoe_threshold_v2(&rep_engine, &id_rep, 100, 2000);
-        assert_eq!(t01, Hash::with_first_u32(0x0ccccccc));
-        assert_eq!((p01 * 1_000_000_f64).round() as i128, 50000);
-        // Inactive identity with 0 reputation
-        let (t02, p02) = calculate_reppoe_threshold_v2(&rep_engine, &id_no_active, 100, 2000);
-        assert_eq!(t02, Hash::with_first_u32(0x00068d8d));
-        assert_eq!((p02 * 1_000_000_f64).round() as i128, 100);
-
-        // Repeat test with 101 witnesses
-        // Active identity with 1% of the reputation
-        let (t01, p01) = calculate_reppoe_threshold_v2(&rep_engine, &id_rep, 101, 2000);
-        assert_eq!(t01, Hash::with_first_u32(0xffffffff));
-        assert_eq!((p01 * 100_f64).round() as i128, 100);
-        // Inactive identity with 0 reputation
-        let (t02, p02) = calculate_reppoe_threshold_v2(&rep_engine, &id_no_active, 101, 2000);
-        assert_eq!(t02, Hash::with_first_u32(0xffffffff));
-        assert_eq!((p02 * 100_f64).round() as i128, 100);
-    }
-
-    // FIXME: Allow for now, wait for https://github.com/rust-lang/rust/issues/67058 to reach stable
-    #[allow(clippy::cast_possible_truncation)]
-    #[test]
-    fn reppoe_100_reputed_nodes_v3() {
-        // Test the behavior of the algorithm when 100 nodes have equally large reputation
-        let mut rep_engine = ReputationEngine::new(1000);
-        let gen_id = |i| PublicKeyHash::from_bytes(&[i; 20]).unwrap();
-        let id_rep = gen_id(0);
-        let id_no_active = gen_id(255);
-        rep_engine.ars_mut().push_activity((0..100).map(gen_id));
-        rep_engine
-            .trs_mut()
-            .gain(
-                Alpha(10),
-                (0..100).map(|i| {
-                    let id = gen_id(i);
-                    (id, Reputation(10_000))
-                }),
-            )
-            .unwrap();
-
-        // Active identity with 1% of the reputation
-        let (t01, p01) = calculate_reppoe_threshold_v3(&rep_engine, &id_rep, 1, 2000);
-        assert_eq!(t01, Hash::with_first_u32(0x028f5c28));
-        assert_eq!((p01 * 1_000_000_f64).round() as i128, 10_000);
-        // Inactive identity with 0 reputation
-        let (t02, p02) = calculate_reppoe_threshold_v3(&rep_engine, &id_no_active, 1, 2000);
-        assert_eq!(t02, Hash::with_first_u32(0x000010c6));
-        assert_eq!((p02 * 1_000_000_f64).round() as i128, 1);
-
-        // Repeat test with 2 witnesses
-        // Active identity with 1% of the reputation
-        let (t01, p01) = calculate_reppoe_threshold_v3(&rep_engine, &id_rep, 2, 2000);
-        assert_eq!(t01, Hash::with_first_u32(0x051eb851));
-        assert_eq!((p01 * 1_000_000_f64).round() as i128, 20_000);
-        // Inactive identity with 0 reputation
-        let (t02, p02) = calculate_reppoe_threshold_v3(&rep_engine, &id_no_active, 2, 2000);
-        assert_eq!(t02, Hash::with_first_u32(0x0000218d));
-        assert_eq!((p02 * 1_000_000_f64).round() as i128, 2);
-
-        // Repeat test with 100 witnesses
-        // Active identity with 1% of the reputation
-        let (t01, p01) = calculate_reppoe_threshold_v3(&rep_engine, &id_rep, 100, 2000);
-        assert_eq!(t01, Hash::with_first_u32(0xFFFF_FFFF));
-        assert_eq!((p01 * 100_f64).round() as i128, 100);
-        // Inactive identity with 0 reputation
-        let (t02, p02) = calculate_reppoe_threshold_v3(&rep_engine, &id_no_active, 100, 2000);
-        assert_eq!(t02, Hash::with_first_u32(0x00068d8d));
-        assert_eq!((p02 * 1_000_000_f64).round() as i128, 100);
-
-        // Repeat test with 100 witnesses
-        // Active identity with 1% of the reputation
-        let (t01, p01) = calculate_reppoe_threshold_v3(&rep_engine, &id_rep, 101, 2000);
-        assert_eq!(t01, Hash::with_first_u32(0xFFFF_FFFF));
-        assert_eq!((p01 * 100_f64).round() as i128, 100);
-        // Inactive identity with 0 reputation
-        let (t02, p02) = calculate_reppoe_threshold_v3(&rep_engine, &id_no_active, 101, 2000);
-        assert_eq!(t02, Hash::with_first_u32(0x00069e54));
-        assert_eq!((p02 * 1_000_000_f64).round() as i128, 101);
-    }
-
-    #[test]
-    // FIXME: Allow for now, since there is no safe cast function from a usize to float yet
-    #[allow(clippy::cast_possible_truncation)]
-    fn reppoe_worst_case() {
-        // Check the worst case reppoe probability in mainnet with an empty ARS, a data request with
-        // the maximum number of witnesses, and taking into account the extra commit rounds.
-
-        fn max_num_witnesses_for_dr_weigth(max_dr_weight: u32) -> u16 {
-            let mut dr = DRTransaction::default();
-            let mut num_witnesses = 1;
-
-            while dr.weight() < max_dr_weight {
-                dr.body.dr_output.witnesses = num_witnesses;
-                num_witnesses += 1;
-            }
-
-            num_witnesses - 1
-        }
-
-        // TODO: copied from chain.rs, make public instead?
-        fn calculate_backup_witnesses(witnesses: u16, commit_round: u16) -> u16 {
-            let exponent = u32::from(commit_round).saturating_sub(1);
-            let coefficient = 2_u16.saturating_pow(exponent);
-
-            witnesses.saturating_mul(coefficient) / 2
-        }
-
-        let consensus_constants = witnet_config::config::consensus_constants_from_partial(
-            &witnet_data_structures::chain::PartialConsensusConstants::default(),
-            &witnet_config::defaults::Mainnet,
-        );
-        let max_dr_weight = consensus_constants.max_dr_weight;
-        // Need to use extra_rounds + 1 because this variable represents the additional rounds
-        let max_commit_rounds = consensus_constants.extra_rounds + 1;
-        let minimum_difficulty = consensus_constants.minimum_difficulty;
-
-        let max_witnesses = max_num_witnesses_for_dr_weigth(max_dr_weight);
-        let max_backup_witnesses = calculate_backup_witnesses(max_witnesses, max_commit_rounds);
-
-        assert_eq!(max_witnesses, 126);
-        assert_eq!(max_witnesses + max_backup_witnesses, 630);
-
-        // Empty ARS
-        let rep_engine = ReputationEngine::new(1000);
-        let id1 = PublicKeyHash::from_bytes(&[1; 20]).unwrap();
-
-        let (_t00, p00) = calculate_reppoe_threshold_v3(
-            &rep_engine,
-            &id1,
-            max_witnesses + max_backup_witnesses,
-            minimum_difficulty,
-        );
-
-        assert_eq!((p00 * 100_f64).round() as i128, 32);
     }
 }
